Handle pkg_path option given by distutils.
Up to now, we assumed that in the case of in-place build, we should
'install' things (extensions, generated headers, etc...) in the src_dir,
which is the same directory as where the scons scripts are. This is not
always true, so we added an option pkg_path to add_sconscript function
in Configuration class (from numpy.distutils) to handle this relatively
rare case (but necessary in matplotlib).